Restructure GraphQL resolvers folder
