+++ src/resolvers/quiz/quiz.types.ts
